# Backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Fitness Tracker API
  version: 1.0.0
  description: |
    Lean Go backend for the core fitness tracking loop: users, exercise library,
    workouts, workout exercises, workout sets, meals, and weight entries.
servers:
  - url: https://localhost:8080
...
paths:
...
  /v1/foods:
    get:
      tags: [Foods]
      summary: List foods
      operationId: listFoods
      parameters:
        - name: name
          in: query
          schema:
            type: string
        - name: brand
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: source
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of foods.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedFoods"
    post:
      tags: [Foods]
      summary: Create food
      operationId: createFood
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFoodRequest"
      responses:
        "201":
          description: Created food.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Food"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
...
  /v1/foods/{id}:
    get:
      tags: [Foods]
      summary: Get food detail
      operationId: getFood
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Food detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Food"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      tags: [Foods]
      summary: Update food
      operationId: updateFood
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFoodRequest"
      responses:
        "200":
          description: Updated food.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Food"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Foods]
      summary: Delete food
      operationId: deleteFood
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "204":
          description: Food deleted.
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          description: Food is in use and cannot be deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/responses/Error"
...
  /v1/meals/{id}/foods:
    get:
      tags: [Foods]
      summary: List foods in a meal
      operationId: listMealFoods
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: List of foods in the meal.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedMealFoods"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    post:
      tags: [Foods]
      summary: Add food to a specific meal
      operationId: addFoodToMeal
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMealFoodRequest"
      responses:
        "201":
          description: Food added to meal.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealFood"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
  /v1/admin/foods:
    get:
      tags: [Admin Dashboard]
      summary: List foods with admin metadata
      operationId: getAdminFoods
      security:
        - BearerAuth: []
      parameters:
        - name: name
          in: query
          schema:
            type: string
        - name: brand
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: source
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of foods with admin-only usage metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAdminFoods"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
...
components:
  parameters:
    ID:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [title, status, error]
      properties:
        type:
          type: string
          description: Problem type URI. Validation errors use a dedicated type; generic errors use `about:blank`.
          example: about:blank
        title:
          type: string
          description: Short, human-readable summary of the error category.
          example: Bad Request
        status:
          type: integer
          description: HTTP status code for the error response.
          example: 400
        detail:
          type: string
          description: Human-readable explanation of the specific error instance.
          example: invalid email or password
        error:
          type: string
          description: Backward-compatible alias of `detail`.
          example: invalid email or password
        errors:
          type: object
          description: Field-specific validation messages when the request fails validation.
          additionalProperties:
            type: string
          example:
            email: email is required
            password: password must be at least 8 characters
...
    Food:
      type: object
      required: [id, name, serving_unit, calories]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        brand:
          type: string
        category:
          type: string
        source:
          type: string
        serving_size:
          type: number
        serving_unit:
          type: string
        calories:
          type: number
        protein:
          type: number
        carbohydrates:
          type: number
        fat:
          type: number
        fiber:
          type: number
        sugar:
          type: number
        sodium:
          type: number
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateFoodRequest:
      type: object
      required: [name, serving_unit, calories]
      properties:
        name:
          type: string
        brand:
          type: string
        serving_size:
          type: number
        serving_unit:
          type: string
        calories:
          type: number
        protein:
          type: number
        carbohydrates:
          type: number
        fat:
          type: number
        fiber:
          type: number
        sugar:
          type: number
        sodium:
          type: number
    UpdateFoodRequest:
      type: object
      properties:
        name:
          type: string
        brand:
          type: string
        serving_size:
          type: number
        serving_unit:
          type: string
        calories:
          type: number
        protein:
          type: number
        carbohydrates:
          type: number
        fat:
          type: number
        fiber:
          type: number
        sugar:
          type: number
        sodium:
          type: number
    MealFood:
      type: object
      required: [id, meal_id, food_id, quantity]
      properties:
        id:
          type: string
          format: uuid
        meal_id:
          type: string
          format: uuid
        food_id:
          type: string
          format: uuid
        quantity:
          type: number
        food:
          $ref: "#/components/schemas/Food"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateMealFoodRequest:
      type: object
      required: [food_id, quantity]
      properties:
        meal_id:
          type: string
          format: uuid
        food_id:
          type: string
          format: uuid
        quantity:
          type: number
...
    PaginationMetadata:
      type: object
      required: [page, limit, total_count, total_pages, has_next]
      properties:
        page:
          type: integer
          description: Current page number.
        limit:
          type: integer
          description: Maximum number of records requested per page.
        total_count:
          type: integer
          description: Total number of records matching the query.
        total_pages:
          type: integer
          description: Total number of pages available for the current limit.
        has_next:
          type: boolean
          description: Whether another page of results exists after the current page.
...
    PaginatedFoods:
      type: object
      required: [data, metadata]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Food"
        metadata:
          $ref: "#/components/schemas/PaginationMetadata"
    AdminFood:
      allOf:
        - $ref: "#/components/schemas/Food"
        - type: object
          required: [meal_usage_count, can_delete]
          properties:
            meal_usage_count:
              type: integer
              format: int64
            can_delete:
              type: boolean
    PaginatedAdminFoods:
      type: object
      required: [data, metadata]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AdminFood"
        metadata:
          $ref: "#/components/schemas/PaginationMetadata"
...
    PaginatedMealFoods:
      type: object
      required: [data, metadata]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/MealFood"
        metadata:
          $ref: "#/components/schemas/PaginationMetadata"
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT